3.8.9 Merge part 3

+++ administrator/components/com_contenthistory/Model/HistoryModel.php
@@ -134,6 +134,12 @@ class HistoryModel extends ListModel
 		{
 			if ($table->load($pk))
 			{
+				if ($table->keep_forever === "1")
+				{
+					unset($pks[$i]);
+					continue;
+				}
+
 				if ($this->canEdit($table))
 				{
 					if (!$table->delete($pk))
